Initial version of Jesher - with a threadpool module
The threadpool is a special one, any thread may start others
and may specify which threads should be started when specific ones are
completed.

The purpose is:
- split current work across threads
- when the thread is completed, start other thread(s)
so it can be a map/reduce algo if needed, or just a regular pool.

No tests yet - concept is similar to my Python implementation in:
https://github.com/LA-Toth/dewi_utils/blob/54f4fab7bf834562b0421b193dee74e971121fbc/dewi_utils/threading.py