git: gerrit: add RelatedChange and RelatedChangeFetcher
The added implementation is faster but limited compared to the features
in: com.urswolfer.gerrit.client.rest:gerrit-rest-java-client.

The goal is to provide a sorted change number/revision number/some other detail
list of the commits of a chain. So from the related changes
the list of [change, parent of change, grandparent of change, ...] is
created.